Update prescription UI: add Total Medicen, improve logic
Replaced order/billing labels with Total Medicen, updated DataGridView styles, and added event handlers for UI elements. Enhanced send-to-cashier logic to require at least one dispensed item, improved search/filtering, and implemented cancel order functionality. Refactored status update methods and made minor UI/UX improvements.

+++ Clinic_DataAccess/clsPrescriptionData.cs
@@ -453,7 +453,7 @@ namespace Clinic_DataAccess
             }
         }
 
-        public static bool UpdatePrescriptionStatus(int prescriptionId, int newStatus)
+        public static bool UpdatePrescriptionStatus(int prescriptionId, byte newStatus)
         {
             try
             {
